feat(hooks): automate deploy and remove hooks
If the same CloudFormation stack also creates the Route53 HostedZone, the plugin currently cannot work. This is because if you would first deploy the stack, the plugin returns an error in `certificateSummary`, because there is no certificate, and if you would first run `create-cert` there is no HostedZone to create the certificate for. The same goes for the removal process.

With this change, the certificate gets created after the HostedZone is deployed, but before the `certificateSummary` is ran. Upon removal, the certificate is removed before the HostedZone is removed.